#0000 - dto model refactor
