+++ src/main/java/si/feri/itk/projectmanager/controller/OccupancyController.java
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import si.feri.itk.projectmanager.dto.request.CreateOccupancyRequest;
-import si.feri.itk.projectmanager.dto.request.UpdateOccupancyRequest;
+import si.feri.itk.projectmanager.dto.request.occupancy.CreateOccupancyRequest;
+import si.feri.itk.projectmanager.dto.request.occupancy.UpdateOccupancyRequest;
 import si.feri.itk.projectmanager.dto.response.ResourceCreatedResponse;
 import si.feri.itk.projectmanager.dto.response.occupancy.CreateOccupancyResponse;
 import si.feri.itk.projectmanager.service.OccupancyService;
