+++ src/main/java/si/feri/itk/projectmanager/controller/ProjectController.java
@@ -15,18 +15,18 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import si.feri.itk.projectmanager.dto.model.PersonDto;
 import si.feri.itk.projectmanager.dto.model.ProjectDto;
-import si.feri.itk.projectmanager.dto.request.AddPersonToProjectRequest;
-import si.feri.itk.projectmanager.dto.request.CreateProjectRequest;
-import si.feri.itk.projectmanager.dto.request.ProjectListSearchParams;
-import si.feri.itk.projectmanager.dto.response.GetPeopleResponse;
+import si.feri.itk.projectmanager.dto.request.project.AddPersonToProjectRequest;
+import si.feri.itk.projectmanager.dto.request.project.CreateProjectRequest;
+import si.feri.itk.projectmanager.dto.request.project.ProjectListSearchParams;
+import si.feri.itk.projectmanager.dto.response.person.GetPeopleResponse;
 import si.feri.itk.projectmanager.dto.response.project.GetProjectResponse;
 import si.feri.itk.projectmanager.dto.response.project.ListProjectResponse;
 import si.feri.itk.projectmanager.dto.response.project.ProjectListStatusResponse;
 import si.feri.itk.projectmanager.dto.response.ResourceCreatedResponse;
-import si.feri.itk.projectmanager.dto.response.project.UpdateProjectRequest;
+import si.feri.itk.projectmanager.dto.request.project.UpdateProjectRequest;
 import si.feri.itk.projectmanager.dto.response.project.UpdateProjectResponse;
 import si.feri.itk.projectmanager.dto.response.statistics.ProjectStatisticsResponse;
-import si.feri.itk.projectmanager.dto.sortinforequest.ProjectSortInfoRequest;
+import si.feri.itk.projectmanager.dto.request.project.ProjectSortInfoRequest;
 import si.feri.itk.projectmanager.paging.request.PageInfoRequest;
 import si.feri.itk.projectmanager.service.PersonService;
 import si.feri.itk.projectmanager.service.ProjectService;
