+++ src/main/java/si/feri/itk/projectmanager/controller/WorkPackageController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import si.feri.itk.projectmanager.dto.request.CreateWorkPackageRequest;
-import si.feri.itk.projectmanager.dto.request.UpdateWorkPackageRequest;
+import si.feri.itk.projectmanager.dto.request.workpackage.CreateWorkPackageRequest;
+import si.feri.itk.projectmanager.dto.request.workpackage.UpdateWorkPackageRequest;
 import si.feri.itk.projectmanager.dto.response.ResourceCreatedResponse;
 import si.feri.itk.projectmanager.service.WorkPackageService;
