+++ src/main/java/si/feri/itk/projectmanager/dto/request/occupancy/UpdateOccupancyRequest.java
@@ -1,21 +1,20 @@
-package si.feri.itk.projectmanager.dto.request;
-
+package si.feri.itk.projectmanager.dto.request.occupancy;
 
 import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.math.BigDecimal;
-import java.time.LocalDate;
+import java.util.UUID;
 
 @Getter
 @Setter
-public class AddPersonToTaskRequest {
-    @Deprecated
+@NoArgsConstructor
+public class UpdateOccupancyRequest {
+    private UUID occupancyId;
     @DecimalMax(value = "1.000")
     @DecimalMin(value = "0.001")
-    private BigDecimal occupancy;
-    private LocalDate startDate;
-    private LocalDate endDate;
+    private BigDecimal value;
 }
