+++ src/main/java/si/feri/itk/projectmanager/dto/request/project/AddPersonToProjectRequest.java
@@ -1,4 +1,4 @@
-package si.feri.itk.projectmanager.dto.request;
+package si.feri.itk.projectmanager.dto.request.project;
 
 import lombok.Getter;
 import lombok.Setter;
