+++ src/main/java/si/feri/itk/projectmanager/dto/request/task/AddPersonToTaskRequest.java
@@ -1,20 +1,21 @@
-package si.feri.itk.projectmanager.dto.request;
+package si.feri.itk.projectmanager.dto.request.task;
+
 
 import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.math.BigDecimal;
-import java.util.UUID;
+import java.time.LocalDate;
 
 @Getter
 @Setter
-@NoArgsConstructor
-public class UpdateOccupancyRequest {
-    private UUID occupancyId;
+public class AddPersonToTaskRequest {
+    @Deprecated
     @DecimalMax(value = "1.000")
     @DecimalMin(value = "0.001")
-    private BigDecimal value;
+    private BigDecimal occupancy;
+    private LocalDate startDate;
+    private LocalDate endDate;
 }
