+++ src/main/java/si/feri/itk/projectmanager/repository/projectlist/CustomProjectListRepo.java
@@ -2,7 +2,7 @@ package si.feri.itk.projectmanager.repository.projectlist;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import si.feri.itk.projectmanager.dto.request.ProjectListSearchParams;
+import si.feri.itk.projectmanager.dto.request.project.ProjectListSearchParams;
 import si.feri.itk.projectmanager.model.ProjectList;
 
 public interface CustomProjectListRepo {
