+++ src/main/java/si/feri/itk/projectmanager/repository/projectlist/CustomProjectListRepoImpl.java
@@ -7,7 +7,7 @@ import com.querydsl.jpa.impl.JPAQuery;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.support.Querydsl;
-import si.feri.itk.projectmanager.dto.request.ProjectListSearchParams;
+import si.feri.itk.projectmanager.dto.request.project.ProjectListSearchParams;
 import si.feri.itk.projectmanager.exceptions.implementation.InternalServerException;
 import si.feri.itk.projectmanager.model.ProjectList;
 import si.feri.itk.projectmanager.model.QProjectList;
