+++ src/main/java/si/feri/itk/projectmanager/service/OccupancyService.java
@@ -5,8 +5,8 @@ import jakarta.transaction.Transactional;
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import si.feri.itk.projectmanager.dto.request.CreateOccupancyRequest;
-import si.feri.itk.projectmanager.dto.request.UpdateOccupancyRequest;
+import si.feri.itk.projectmanager.dto.request.occupancy.CreateOccupancyRequest;
+import si.feri.itk.projectmanager.dto.request.occupancy.UpdateOccupancyRequest;
 import si.feri.itk.projectmanager.dto.response.occupancy.CreateOccupancyResponse;
 import si.feri.itk.projectmanager.dto.response.occupancy.OccupancyWarning;
 import si.feri.itk.projectmanager.exceptions.implementation.IllegalResourceAccess;
