+++ src/main/java/si/feri/itk/projectmanager/service/ProjectService.java
@@ -9,17 +9,16 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import si.feri.itk.projectmanager.dto.common.Duration;
 import si.feri.itk.projectmanager.dto.model.ProjectDto;
-import si.feri.itk.projectmanager.dto.request.AddPersonToProjectRequest;
-import si.feri.itk.projectmanager.dto.request.CreateProjectRequest;
-import si.feri.itk.projectmanager.dto.request.ProjectListSearchParams;
+import si.feri.itk.projectmanager.dto.request.project.AddPersonToProjectRequest;
+import si.feri.itk.projectmanager.dto.request.project.CreateProjectRequest;
+import si.feri.itk.projectmanager.dto.request.project.ProjectListSearchParams;
 import si.feri.itk.projectmanager.dto.response.project.ListProjectResponse;
 import si.feri.itk.projectmanager.dto.response.project.ProjectListStatusResponse;
-import si.feri.itk.projectmanager.dto.response.project.UpdateProjectRequest;
-import si.feri.itk.projectmanager.dto.response.project.UpdateProjectResponse;
+import si.feri.itk.projectmanager.dto.request.project.UpdateProjectRequest;
 import si.feri.itk.projectmanager.dto.response.statistics.PersonWorkDto;
 import si.feri.itk.projectmanager.dto.response.statistics.ProjectMonthDto;
 import si.feri.itk.projectmanager.dto.response.statistics.ProjectStatisticsResponse;
-import si.feri.itk.projectmanager.dto.sortinforequest.ProjectSortInfoRequest;
+import si.feri.itk.projectmanager.dto.request.project.ProjectSortInfoRequest;
 import si.feri.itk.projectmanager.exceptions.implementation.BadRequestException;
 import si.feri.itk.projectmanager.exceptions.implementation.ItemNotFoundException;
 import si.feri.itk.projectmanager.mapper.ProjectMapper;
