+++ src/main/java/si/feri/itk/projectmanager/service/TaskService.java
@@ -4,9 +4,9 @@ import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import si.feri.itk.projectmanager.dto.request.AddPersonToTaskRequest;
-import si.feri.itk.projectmanager.dto.request.CreateTaskRequest;
-import si.feri.itk.projectmanager.dto.request.UpdateTaskRequest;
+import si.feri.itk.projectmanager.dto.request.task.AddPersonToTaskRequest;
+import si.feri.itk.projectmanager.dto.request.task.CreateTaskRequest;
+import si.feri.itk.projectmanager.dto.request.task.UpdateTaskRequest;
 import si.feri.itk.projectmanager.exceptions.implementation.BadRequestException;
 import si.feri.itk.projectmanager.exceptions.implementation.ItemNotFoundException;
 import si.feri.itk.projectmanager.model.PersonOnTask;
