+++ src/main/java/si/feri/itk/projectmanager/service/WorkPackageService.java
@@ -6,8 +6,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import si.feri.itk.projectmanager.dto.common.Duration;
 import si.feri.itk.projectmanager.dto.common.IDuration;
-import si.feri.itk.projectmanager.dto.request.CreateWorkPackageRequest;
-import si.feri.itk.projectmanager.dto.request.UpdateWorkPackageRequest;
+import si.feri.itk.projectmanager.dto.request.workpackage.CreateWorkPackageRequest;
+import si.feri.itk.projectmanager.dto.request.workpackage.UpdateWorkPackageRequest;
 import si.feri.itk.projectmanager.exceptions.implementation.ItemNotFoundException;
 import si.feri.itk.projectmanager.model.Project;
 import si.feri.itk.projectmanager.model.WorkPackage;
