+++ src/test/java/si/feri/itk/projectmanager/util/service/ProjectServiceUtilTest.java
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
-import si.feri.itk.projectmanager.dto.request.AddPersonToProjectRequest;
-import si.feri.itk.projectmanager.dto.request.CreateProjectRequest;
+import si.feri.itk.projectmanager.dto.request.project.AddPersonToProjectRequest;
+import si.feri.itk.projectmanager.dto.request.project.CreateProjectRequest;
 import si.feri.itk.projectmanager.exceptions.CustomRuntimeException;
 import si.feri.itk.projectmanager.model.Project;
 import si.feri.itk.projectmanager.model.person.Person;
