+++ src/test/java/si/feri/itk/projectmanager/util/service/TaskServiceUtilTest.java
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
-import si.feri.itk.projectmanager.dto.request.AddPersonToTaskRequest;
-import si.feri.itk.projectmanager.dto.request.CreateTaskRequest;
+import si.feri.itk.projectmanager.dto.request.task.AddPersonToTaskRequest;
+import si.feri.itk.projectmanager.dto.request.task.CreateTaskRequest;
 import si.feri.itk.projectmanager.exceptions.CustomRuntimeException;
 import si.feri.itk.projectmanager.model.PersonOnTask;
 import si.feri.itk.projectmanager.model.Task;
